+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.1 (work in progress)
 
 - Add `GitHubAPI` to get licenses from GitHub API
-- Update `AcknowListViewController` to get missing licenses from GitHub API, with new `canFetchLicenseFromGitHub` property to disable this behavior
+- Update `AcknowListViewController` and `AcknowListSwiftUIView` to get missing licenses from GitHub API, with new `canFetchLicenseFromGitHub` property to disable this behavior
 
 
 ## 3.0.1 (24 November 2022)